\name{atc-deprecated}
\alias{atc-deprecated}
\title{Deprecated functions in package \sQuote{atc}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{atc} only, and will be defunct in one of the next releases.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{atcData: \code{\link{atcs}}}

  }
}
