% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{atcLevel}
\alias{atcLevel}
\alias{toAtcLevel}
\title{Utilities to work with ATC codes}
\usage{
atcLevel(x)

toAtcLevel(x, level = 1)
}
\arguments{
\item{x}{For \code{atcLevel} and \code{toAtcLevel}: \code{character}
representing ATC code(s).}
}
\value{
For \code{atcLevel}: an \code{integer} of length 1 representing the
    ATC level. The function returns \code{NA} if the length of the ATC code
    does not match any of the levels (i.e. the length of the code is not
    equal to 1, 3, 4, 5 or 7).

\code{toAtcLevel}: if \code{length(level) == 1}: returns a
    \code{character} of the same length then \code{x}. If
    \code{length(level) > 1}: returns a \code{matrix} of ATC codes, each
    column corresponding to elements in \code{x}, each row to ATC levels.
}
\description{
\code{atcLevel} determines the ATC \emph{level} for the
    provided ATC code. Note that the level is purely determined by the
    number of characters. The function does not evaluate whether the
    provided key represents a valid ATC key/code.
    See \code{\link{AtcDb}} for an overview of the various levels.

\code{toAtcLevel} transforms the provided ATC code(s) to a lower
    level code. The function throws an error if the provided code has a lower
    level than the level specified with parameter \code{level}.
}
\examples{

## Get the level for given ATC codes
atcLevel(c("C07AB07", "A12AX"))

## Invalid ATC code
atcLevel("AA")

## Transform an ATC code to a lower level.
toAtcLevel("A01", level = 1)

## Transform one ATC code into two different levels
toAtcLevel("A01CD04", level = c(1, 3))

## Transform two codes.
toAtcLevel(c("A01CD", "A01CD04"), level = 1)

## Transform two codes in 3 levels
toAtcLevel(c("A01CD", "A01CD04"), level = c(1, 3, 2))
}
\seealso{
\code{\link{AtcDb}} for a listing and description of ATC levels.
}
\author{
Johannes Rainer
}
