% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClass.R
\name{findClass}
\alias{findClass}
\title{findClass}
\usage{
findClass(medname, tier = 4)
}
\arguments{
\item{the}{raw medication that user is interested in}

\item{the}{tier that user is interested in. Default to 4.}
}
\value{
two lists: the first list contains the table that maps the medication to its corresponding drugs; the second list contains the tables that map drugs to its ATC
}
\description{
This function returns the ATC of the drugs in the medication.
}
