% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_atc_code.R
\name{search_atc_code}
\alias{search_atc_code}
\title{Retrieve Classification Details for a Anatomical Therapeutic Chemical (ATC) Code}
\usage{
search_atc_code(ATC_CODE)
}
\arguments{
\item{ATC_CODE}{a character string of a valid ATC code}
}
\description{
Retrieve the hierarchical classification details for a Anatomical Therapeutic Chemical (ATC) code
from the World Health Organisation (WHO) Collaborating Centre for Drug Statistics Methodology website (https://www.whocc.no)
}
